some warnings fixed
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@16548 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ desktop/geogebra/euclidian/EuclidianStatic.java
@@ -57,7 +57,7 @@ public class EuclidianStatic extends geogebra.common.euclidian.EuclidianStatic{
 						elements[i], font, ((GeoText) geo).isSerifFont(),
 						fgColor, bgColor, false);
 
-				int height = (int) dim.getHeight();
+				int height = dim.getHeight();
 
 				// depth += dim.depth;
